+++ demisto_sdk/__main__.py
@@ -251,9 +251,9 @@ def unify(**kwargs):
          'This applies only when the -g flag is supplied.'
 )
 @click.option(
-    '-iu', '--ignore-untracked',
+    '-iu', '--include-untracked',
     is_flag=True,
-    help='Whether to ignore untracked files in the validation.'
+    help='Whether to include untracked files in the validation.'
 )
 @click.option(
     '-a', '--validate-all', is_flag=True, show_default=True, default=False,
@@ -326,7 +326,7 @@ def validate(config, **kwargs):
             json_file_path=kwargs.get('json_file'),
             skip_schema_check=kwargs.get('skip_schema_check'),
             debug_git=kwargs.get('debug_git'),
-            ignore_untracked=kwargs.get('ignore_untracked')
+            include_untracked=kwargs.get('include_untracked')
         )
         return validator.run_validation()
     except (git.InvalidGitRepositoryError, git.NoSuchPathError, FileNotFoundError) as e: